Moved some ai helpers to their own file so they can be used in other places if needed. Adjusted some documentation.
